% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pv_visualise.R
\name{createTable2}
\alias{createTable2}
\title{createTable2}
\usage{
createTable2(q, format = "long")
}
\arguments{
\item{q}{vector of features e.g. c("q1","q3")}

\item{format}{return the table in long or wide format}
}
\value{
n x m contingency table
}
\description{
creates a contingency table for a pair of features from pv_data
}
