% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PVsurvey.R
\name{dbeta_util}
\alias{dbeta_util}
\title{dbeta_util}
\usage{
dbeta_util(x, shape1, shape2)
}
\arguments{
\item{x}{in -1,1}

\item{shape1}{"a" shape parameter}

\item{shape2}{"b" shape parameter}
}
\value{
function value
}
\description{
Beta function distribution generalised to interval -1,1
}
