% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pv_visualise.R
\name{featureImportance}
\alias{featureImportance}
\title{featureImportance}
\usage{
featureImportance(shap_scores_long, feature_code)
}
\arguments{
\item{shap_scores_long}{shap scores in long format (output of get_shap_scores)}

\item{feature_code}{feature code partial utility to be plotted}
}
\value{
a plot
}
\description{
makes a boxplot of shap score dependences on feature values
}
