% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PVsurvey.R
\name{get_boosted_tree_model}
\alias{get_boosted_tree_model}
\title{ge_boosted_tree_model}
\usage{
get_boosted_tree_model(
  pv_data_in,
  learning_rate = 0.02,
  tree_depth = 5,
  k_crossvalidation = 5,
  complexity_factor = 1,
  s = 0.15,
  epsilon = 0.7
)
}
\arguments{
\item{pv_data_in}{input survey data}

\item{learning_rate}{eta. typical value 0.02}

\item{tree_depth}{tree depth typical value 5}

\item{k_crossvalidation}{k-fold cross validatin typical value 5}

\item{complexity_factor}{"over-fitting" enhancement relative to optimal model complexity from cross-validation. Values in range 1-1.5.}

\item{s}{utility uncertainty (standard deviation) default 0.15}

\item{epsilon}{survey hypothetical bias (zero bias corresponds to 1) default 0.7}
}
\value{
xgboost model
}
\description{
creates a cross-validated boosted tree regression model from pv survey data
}
