% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PVsurvey.R
\name{get_empirical_partial_utilities}
\alias{get_empirical_partial_utilities}
\title{get_empirical_partial_utilities}
\usage{
get_empirical_partial_utilities(shap_scores_long, stat = "median")
}
\arguments{
\item{shap_scores_long}{shap scores (partial utilities)}

\item{stat}{median (default) or mean}
}
\value{
dataframe
}
\description{
partial (dis) utilities for pv adoption derived from survey
}
