% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PVsurvey.R
\name{get_model_weights}
\alias{get_model_weights}
\title{get_model_weights}
\usage{
get_model_weights(shap_scores_long, stat = "median")
}
\arguments{
\item{shap_scores_long}{shaps scores (partial utilities)}

\item{stat}{median (default) or mean}
}
\value{
a dataframe with colums ID w_q9_1  w_qsp21 W_theta
}
\description{
The agent weights for financial, social and barrier terms
}
