% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PVsurvey.R
\name{get_shap_scores}
\alias{get_shap_scores}
\title{get_shap_scores}
\usage{
get_shap_scores(pv_data_in, bst)
}
\arguments{
\item{pv_data_in}{input survey data (training). Use pv_data for full survey data or pv_data_oo for owner-occupiers}

\item{bst}{boosted tree model from xgboost}
}
\value{
shap scores for all agents and features, in long format and including BIAS (same for all agents)
}
\description{
returns partial utilities (shap scores) for each agent
}
