% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PVsurvey.R
\name{map_likertscores_to_utilities}
\alias{map_likertscores_to_utilities}
\title{map_likertscores_to_utilities}
\usage{
map_likertscores_to_utilities(s = 0.15, epsilon = 0.75)
}
\arguments{
\item{s}{agent utility uncertainty (standard deviation of utility at time of survey)}

\item{epsilon}{probability scale factor (to allow for hypothetical bias) 1= no hypothetical bias. default 0.75.}
}
\value{
a dataframe of shape parameters and expected utility values corresponging to likert probabilities
}
\description{
maps likert scores 1..5 to adoption probabilities and then to expected utility.
Epsilon controls the degree and sign of the hypothetical bias correction needed when tuning the ABM
}
