% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PVsurvey.R
\name{pbeta_util}
\alias{pbeta_util}
\title{pbeta_util}
\usage{
pbeta_util(x, shape1, shape2)
}
\arguments{
\item{x}{real in range -1 to 1}

\item{shape1}{"a" shape parameter}

\item{shape2}{"b" shape parameter}
}
\value{
generalised beta function value
}
\description{
Beta function probability generalised to range -1,1
}
