% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pv_visualise.R
\name{plot_weights}
\alias{plot_weights}
\title{plot_weights}
\usage{
plot_weights(shap_scores_long, clip = 0.05)
}
\arguments{
\item{shap_scores_long}{partial utilities}

\item{clip}{fraction of observations to clip (i.e. reset to clip/2, 1-clip/2 quantile values). thetas are unclipped.}
}
\value{
a pairs plot
}
\description{
plot_weights
}
