% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PVsurvey.R
\name{probs_from_shape_params}
\alias{probs_from_shape_params}
\title{probs_from_shape_params}
\usage{
probs_from_shape_params(s)
}
\arguments{
\item{s}{standard deviation of Beta distribution}
}
\value{
dataframe
}
\description{
mapping between shape parameters of generalised Beta function and total probability that value > 0
}
