% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PVsurvey.R
\name{shape_params_from_prob}
\alias{shape_params_from_prob}
\title{shape_params_from_prob}
\usage{
shape_params_from_prob(df.in, prob)
}
\arguments{
\item{df.in}{dataframe produced by probs_from_shape_params()}

\item{prob}{value in range 0,1}
}
\value{
shape paramaters (a and b)
}
\description{
returns shape parameters corresponding to an adoption probability (i.e. area of generalised beta distribution > 0)
}
