% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PVsurvey.R
\name{transform_to_utils}
\alias{transform_to_utils}
\title{transform_to_utils}
\usage{
transform_to_utils(pv_data, s = 0.15, epsilon = 0.7)
}
\arguments{
\item{pv_data}{survey data including qsp22_7 (Likert scores for adopting pv)}

\item{s}{utility uncertainty (standard deviation) default 0.15}

\item{epsilon}{survey hypothetical bias (zero bias corresponds to 1) default 0.7}
}
\value{
modified pv survey dataframe
}
\description{
Transforms survey Likert scores for likelihood of pv adoption (1..5) to utilities in range -1,1.
The transformation depends on s (utility uncertainty) and epsilon (survey hypothetical bias estimate).
}
