% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PVsurvey.R
\name{find_optimum_rounds_from_crossvalidation}
\alias{find_optimum_rounds_from_crossvalidation}
\title{find_optimum_rounds_from_crossvalidation}
\usage{
find_optimum_rounds_from_crossvalidation(
  pv_data,
  learning_rate = 0.02,
  tree_depth = 5,
  k_crossvalidation = 5
)
}
\arguments{
\item{pv_data}{solar PV data}

\item{learning_rate}{eta parameter (default 0.02)}

\item{tree_depth}{maximum tree depth (default 5)}

\item{k_crossvalidation}{k-fold cross validation}
}
\value{
n_opt
}
\description{
helper function to find optimum learning complexity for for given learning rate and tree depth
}
